Complete Task 4 of Flex
Centered information items in a flex on the center of the space, so text
and image are centered horizontally.

Corrections from Solutoin:
When trying to align text on all elements you can just add that to the
body. And you don't have to justify content, wrap or even use flex on
flex items because you can use max-width to wrap the text under the
image.

[Ticket: X]