Complete Task 6 of Flexbox
Learned how to position header (navigation), content and footer into a
nice pile using flexbox. Also was reminded that I can center the content
of the page vertically by using flex: 1, function where the height and
width are respected and it would shrink and grow flexibly while keeping
the footer and header in place at the top and bottom of page.

[Ticket: X]